requirements.txt
alembic==1.11.1
anyio==3.7.0
asyncpg==0.27.0
cfgv==3.3.1
click==8.1.3
distlib==0.3.7
exceptiongroup==1.1.1
fastapi==0.97.0
filelock==3.12.2
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
identify==2.5.26
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
nodeenv==1.8.0
platformdirs==3.10.0
pre-commit==3.3.3
psycopg2-binary==2.9.6
pydantic==1.10.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starlette==0.27.0
strawberry-graphql==0.187.1
typing_extensions==4.6.3
uvicorn==0.22.0
uvloop==0.17.0
virtualenv==20.24.2
watchfiles==0.19.0
websockets==11.0.3
